Rename `utils` folders to more generalizable names
I ran into an issue in which the folder names I had used within the `website.utils` folder were too specific, meaning that some potential utility classes would not fit under the current schema despite being very related to the current utility classes. I've changed this by changing the names to the more general names of `db_fields` and `form` (since `forms` caused an app label issue with Django's `forms`).
