+++ website/utils/__init__.py
@@ -1,4 +1,4 @@
-from website.utils.fields import PublicationDateTimeField, ResizedHashNameImageField
-from website.utils.widgets import ClearableImageInput
+from website.utils.db_fields import PublicationDateTimeField, ResizedHashNameImageField
+from website.utils.form import ClearableImageInput
 
 __all__ = ["ClearableImageInput", "PublicationDateTimeField", "ResizedHashNameImageField"]
