+++ website/utils/form/apps.py
@@ -5,4 +5,4 @@ from django.apps import AppConfig
 class WidgetsConfig(AppConfig):
     """Config of the home app."""
 
-    name = 'website.utils.widgets'
+    name = 'website.utils.form'
