+++ website/utils/form/clearable_image_widget.py
